requirements.txt
mysql-connector-python~=8.1.0
requests~=2.31.0
pandas~=2.0.3
beautifulsoup4~=4.12.2
matplotlib~=3.7.2
nltk~=3.8.1
python-dotenv~=1.0.0